Precise movements and adjusting speed

+++ src/com/creatiom/rain/entity/projectile/WizardProjectile.java
@@ -8,7 +8,7 @@ public class WizardProjectile extends Projectile {
 	
 	public static final int FIRE_RATE = 12; // Higher is slower
 
-	public WizardProjectile(int x, int y, double dir) {
+	public WizardProjectile(double x, double y, double dir) {
 		super(x, y, dir);
 		range = 250;
 		speed = 4;
