+++ src/com/creatiom/rain/level/Level.java
@@ -148,14 +148,14 @@ public class Level {
 	public List<Entity> getEntities(Entity e, int radius) {
 		List<Entity> results = new ArrayList<Entity>();
 		
-		int ex = e.getX();
-		int ey = e.getY();
+		int ex = (int)e.getX();
+		int ey = (int)e.getY();
 		
 		for(int i = 0; i < entities.size(); i++) {
 			Entity entity = entities.get(i);
 			
-			int x = entity.getX();
-			int y  = entity.getY();
+			int x = (int)entity.getX();
+			int y  = (int)entity.getY();
 			
 			int dx = Math.abs(x - ex);
 			int dy = Math.abs(y - ey);
@@ -175,14 +175,14 @@ public class Level {
 		
 		List<Player> results = new ArrayList<Player>();
 		
-		int ex = e.getX();
-		int ey = e.getY();
+		int ex = (int)e.getX();
+		int ey = (int)e.getY();
 	
 		for(int i = 0; i < players.size(); i++) {
 			Player player = players.get(i);
 				
-			int x = player.getX();
-			int y = player.getY();
+			int x = (int)player.getX();
+			int y = (int)player.getY();
 				
 			int dx = Math.abs(x - ex);
 			int dy = Math.abs(y - ey);
